Tidied up PhotonOutput and ElectronEnergy
