+++ scripts/Universal.py
@@ -432,7 +432,7 @@ def Create_Legend(	df, color_column, custom_color_boolean,
 	# large enough and otherwise use the large Turbo palette. (Will error out if
 	# Turbo is not large enough.
 	else:
-		if len(color_list) < 8:
+		if (len(color_list) < 8) and (len(color_list) > 2):
 			color_palette = list(Colorblind[len(color_list)])
 		else:
 			# This will throw an error if you have more than 256 unique items
